summarize bins

Returns a summary of all bins created by `bin_cols` in a data frame. Takes no arguments other than the data frame but relies on regular expressions based of the `bin_cols` output in order to identify the corresponding columns.

bin_summary(mdb, ...)

Arguments

mdb
dataframe output from bin_cols

...
optional tidyselect specification for specific cols

Value

a tibble

Examples

iris %>%
bin_cols(Sepal.Width) %>%
bin_summary()

#> # A tibble: 9 × 14
#>   column      method     n_bins .rank  .min .mean  .max .count .uniques relative_value
#>   <chr>       <chr>       <int> <int> <dbl> <dbl> <dbl>  <int>    <int>          <dbl>
#> 1 Sepal.Width equal freq      9     9   3.7  3.89   4.4     15        7          100  
#> 2 Sepal.Width equal freq      9     8   3.5  3.54   3.6     10        2           90.9
#> 3 Sepal.Width equal freq      9     7   3.3  3.37   3.4     18        2           86.5
#> 4 Sepal.Width equal freq      9     6   3.2  3.2    3.2     13        1           82.2
#> 5 Sepal.Width equal freq      9     5   3.1  3.1    3.1     11        1           79.6
#> 6 Sepal.Width equal freq      9     4   2.9  2.97   3       36        2           76.3
#> 7 Sepal.Width equal freq      9     3   2.8  2.8    2.8     14        1           71.9
#> 8 Sepal.Width equal freq      9     2   2.6  2.66   2.7     14        2           68.4
#> 9 Sepal.Width equal freq      9     1   2    2.37   2.5     19        5           60.8
#> # … with 4 more variables: .sum <dbl>, .med <dbl>, .sd <dbl>, width <dbl>