Make Bins

Make bins in a tidy fashion. Adds a column to your data frame containing the integer codes of the specified bins of a certain column. Bin types are input as logical statements. Multiple bin types can be set to `TRUE`, creating multiple types of bins for one column. Specifying multiple columns is only intended for supervised binning, so mutliple columns can be simultaneously binned optimally with respect to a target variable.

make_bins(
  .data,
  col,
  n_bins = 10,
  bin_equal_frequency = F,
  bin_equal_width = F,
  bin_equal_value = F,
  bin_kmeans = F,
  bin_xgboost = F,
  ...,
  target = NULL,
  pretty_labels = TRUE,
  seed = 1
)

Arguments

.data
a data frame

col
a column, vector of columns, or tidyselect

n_bins
number of bins

bin_equal_frequency
logical, bin type equal frequency (quantile, ntile)

bin_equal_width
logical, bin type equal width (cut)

bin_equal_value
logical, bin type equal value

bin_kmeans
logical, bin type kmeans

bin_xgboost
logical, supervised binning with xgboost. target must be specified

...
params to be passed to selected binning method

target
unquoted column for supervised binning

pretty_labels
logical. If true returns interval label rather than integer rank

seed
seed for stochastic binning (xgboost)

Value

a tibble